Made convertToLinkedList much more efficient
The biggest realization here was that each time a tree was converted to a doubly-linked list representing its in-order traversal, the important parts of that linked list were the head and tail. At each node of the tree, the current node needs to be placed at the tail of the linked list of its left subtree and at the head of the linked list of its right subtree. It then needs to pass the information of the current head and tail on to its parent node. If either its left or right child are null, the current node becomes the head or tail of the linked list respectively. This is the base case for the recursion: the head and tail of a doubly-linked list created by a leaf node are both the leaf node.

So to accomplish this, the convertToLinkedList method now returns a "meta-node" where the data is ignored, but the left child of said meta-node points to the head of the linked list and the right child points to the tail.

TODO: Add a README